+++ js/i18n/eden-map/ru.js
@@ -97,7 +97,7 @@ export default {
     subTabMap: 'Карта Эдема',
     subTabLoyalty: 'Верность Эдема',
     subTabBounty: 'Королевская награда Eden X2',
-    subTabSeason: 'Текущий сезон',
-    subTabPrevious: 'Прошлые сезоны',
+    subTabSeason: 'Текущий сезон · {currentSeason}',
+    subTabPrevious: 'Прошлые сезоны · {previousSeason}',
   },
 };